+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   The data transfer module is configured with the `modelardb.transfer` setting.
 - The value of `modelardb.dynamic_split_fraction` is now used directly as set.
 - Renamed `gaps` in the Segment View to `offsets` for both Apache Spark and H2.
+- Renamed `mid` to `mtid` in `ORCStorage` and `ParquetStorage` for consistency.
 
 ### Deprecated
 ### Removed